feat: make run loop handle synchronous operations
When written, the scheduling run loop that drives the executor was
written assuming all operations are asynchronous. This was the correct
design, as a) some operations were already asynch (execute) b) all of
graphnets operations we asynchronous and c) in our future plans to have
the loop run in job-server rather than in the backend, then all
operations would be async anyway.

However, the local executor's prepare() and finalize() are still
synchronous. We made this work by pretending they were async from
a state transition POV, and the next loop iteration would pick up the
correct state. For example, calling prepare() would return the state
PREPARING, even though technically its PREPARED. This worked fine for
a while.

However, when we added telemetry, this sleight of hand meant that all
our timing measurements performed by the loop for synchronous actions
are wrong. For example, we record the state change to PREPARING when we
return from the prepare() call, but actually, all the work (which can
take a long time) has already been done.

Additionally, our live status traces report the wrong current statues
while the synchronous step is going on, leading to inaccurate diagnoses
current load.

We've attempted to fix this in various ways, before it became clear that
to have accurate duration measurements, we need to measure at different
points when the operation is synchronous.

So, this change alters the run loops handling of synchronous operations:

 - if an operation is synchronous, we record the state change time
   *before* calling the executor operation. This provides accurate
   telemetry timings.
 - we change the expectated state to be the completed one (e.g. PREPARED
   or FINALIZED.
 - we change the local executor implementation to return the real state
   to match the above expectation
 - we extend the StubExecutorAPI implementation to implement
   synchronous operations according to the above changes