feat: record the high water mark for cpu/memory for a job
We don't do anything with it right now, but maybe in future we will,
e.g. displaying it to users, or using it to schedule jobs better.

There's a bit of wrinkle here though: we're kinda of breaking the
scheduler/executor layering a bit, by storing these values directly in
the job-runner db.

Ideally, the executor would be responsible for tracking resource usage,
both for live metrics and for high-water-marks. But the current executor
was not designed with this in mind.

So this takes the easy option, and makes the current record_stats thread
handle it. With the caveat that its specific to the local docker
executor.