fix: Fix buggy `assert_new_jobs_created` logic
I'm not sure if this has always been broken, or whether things have
changed elsewhere in the codebase such that assumptions about the
contents of `current_jobs` no longer hold.

In any case, the bug would be triggered when all the actions a user
requested were already running *and* some other action which they didn't
request had already failed.

As the cost of passing in `job_request` we can radically simplify the
logic here.