fix: handle BindMount directory already existing
With the BindMount volume api, if the cleanup fails for some reason,
then the directory is left on disk. If this happens with a job that is
then retried (via retry_job or db maintenance), it will reuse the same
job id and thus the same directory path. If the timestamp file doesn't
exist in that dir (either because it got deleted, or the initial
prepare() didn't complete), then it will attempt to create the volume
again, which failed as it already existed.

This change allows the bind mount dir to already exist, and also logs
any errors with clean up, rather than swallowing them, so if the cleanup
fails again we get some visibility on why.

Motivated by an actual case of this happening:

https://ui.honeycomb.io/bennett-institute-for-applied-data-science/environments/production/datasets/jobrunner/trace/JnsHHM5QpJr?fields[]=c_name&fields[]=c_service.name&span=dbd6d9b58eb310e8