fix: delete as many files as we can
Currently, the bindmount api's delete_volume method errors on TPP once
it errors on one file. This needs fixing, but may take a bit of time.

To handle it better in the meantime, record but don't error on these
files, and log them. That way, we delete all the files we can, and leave
only those we can't.