fix: run as non-root user with BindMountVolumeAPI
Current, delete_volume() doesn't work properly with the
BindMountVolumeAPI. Because the containers are run as root, files they
create are owned by root, and this prevents jobrunner from deleting
them, and we currently manually have to delete them.

This change runs jobs with the jobrunner uid/gid when using
BindMountVolumeAPI, which solves the problem.

I have tested that all our current images work when run as non-root with
some sample studies, so I think this is good to go.