Remove code relating to ancient Data Builder behaviour
There is now no longer a command called `databuilder generate_dataset`.
The subcommand was changed to `generate-dataset` a very long time ago,
and the image itself was renamed to `ehrql` a little while ago now. So
this code has been doing nothing.

This allows us to remove `is_extraction_command` entirely which was a
potential footgun as it looked like the function for deciding whether a
command should have database access, but that wasn't its role.