Remove left-over logic from switch to finalizing everything
ExecutorState.ERROR can be confusing. It is meant to indicate that the
*executor* has errored - a job exiting with an exit code is expected
behaviour, not an *executor* error.

As such, `get_status()` should not be returning ExecutorState.ERROR when
a job is OOM killed - that's business as usual, and should be handled by
`finalise()` as normal.

When we refactored job-runner recently for everything to go via
`finalise()`, I think this behaviour was mistakenly preserved from the
old way of doing things. It resulted in job-runner treating OOM kills
as INTERNAL_ERRORS, and obscuring the message to the users and
needlessly paging teck-support.

Now we only handle the OOM case in one place, I de-factored the
functions that were previously shared between the two locations

Fixes #634