Fix latent bug in manifest repo
If a resuable action happened to be the last job to execute, then the
repo in the manifest file was set to the action's repo, not the study's
repo.

However, we no longer need the repo in the manifest files, so just set
it none to avoid this problem.

Once we have a releases UI, we can get rid of the manifest file code
alltogether.