Fix missing job metrics
We use `str.lstrip("os-job-")` to filter the list of metrics. However,
strip doesnt work like that, a fact that has bitten me many times
previously.

It meant that if a job id started with any of the 'osjb' characters, it
would be stripped from the id in the dict, leadint to it not being
found.

Sadly, we're on 3.8, so we can't use removeprefix, so we fake it.

This was found by the added telemetry in earlier PR, but also needed
some cowboyed additonal bits in prod, to confirm, now undone.

Fixes #634