Move "db" attribute from workspace to job request
This was already how it was modelled internally, so this change is just
about how we translate the JSON we get from Job Server.

There is fallback code to fetch the attribute from `worksapce` if it
isn't where we expect it, so this is safe to deploy before we update Job
Server.