Limit level 4 file size
This adds a global config, `config.MAX_LEVEL4_FILESIZE` (which defaults
to 16mb) which limits the size of files that will be copied to level
4 storage.

The goal here is to prevent accidental copying of files to level 4 that
should not be. Level 4 files should be aggregate data, and potentially
reviewable/releaseable by output checkers. Datasets including
pseudonymized patient level data should not be marked as
`moderately_senstitive`, and this is one thing to prevent them being
done so by mistake.

When triggered, it communciates the fact that files have not been
copied to users in two ways:

1) It includes the list of files not cpied in the job satus_message,
   which will be visible at jobs.opensafely.org

2) It writes file with the same name plus `.txt` file with a message, so
   it discoverable in level 4. It deletes this file if files are
   successfully copied.

Fixes #298