Collect and store per-job aggregate metrics
Stores and updates running totals/aggregates of metrics whilst a job is
EXECUTING.

To keep executor state separate from scheduler state, we store this
state in a separate db file altogether. This prevents lock contention on
the main db file. Unlike the last time we did this, it should not grow
to exessive size, as we only store 1 small row per job.

For each metric, we calculate and store: lastest sample, cumsum, current
mean (measured at this point in time), and peak value.

We also add them to the tick trace spans as can be useful.