Add StatusCode.STALE_CODELISTS
Previously, the handling of a stale codelists meant an INTERNAL_ERROR
was raised, which needlessly alerted tech-support.

This change adds a new StatusCode to track this specific error. This
allows us to to avoid re-using INTERNAL_ERROR, and also make it easier
to measure stale codelist frequency.

In updating the state diagram, I realised we were missing
a CREATED->SUCCEEDED transition, so added that too.