Use requires db logic from latest pipeline release
Previously, we had conflicting definitions of what jobs required db
access. This change uses the canonical definition from the pipeline
library.

This does mean we need to record this in the database when we create the
job, as that's the only time we have the pipeline object available.

Additionally, we used that logic outside the Pipeline model when
validating reusable actions.

Fixes #250