Write output metadata to manifest.json
The main motivation for this was allowing airlock to access commit
data for level 4.

However, it was simple to extend this idea to include a) all outputs and
b) much more metadata, like job, action, size, timestamp, content hash,
and whether a file has been excluded from level 4 and why. This data is
useful for airlock, but also potentially for other systems and general
reporting.