Switch to using --mount instead of --volume
The --mount option is more strict that the --volume option, and will
error if the host source directory does not exist, whereas --volume will
silently create a directory.  We want the strict behaviour, so this
changes to use --mount.

The motivation for this change is to support job-runner running in
docker, which requires that the paths it uses are correct from the
host's perspective.  Currently, a incorrect path will be sort of work,
as the directory would be created inside the container, but only later
would the issue arise.

https://docs.docker.com/engine/storage/bind-mounts/#differences-between--v-and---mount-behavior