Fix docker image permissions
This simplifies the docker container by removing all the previous
user/group management inside the container as it is not needed. Instead,
we used the much simpler solution of adding an additional group to the
running user that is the host's docker group id. This allows the
container to have permission to access the host's bind-mounted
/var/run/docker.sock docker, but the container itself can run as the
local running user, which DTRT with respect to file permissions on bind
mounts.