Fix docker configuration
The job-runner container uses the host's docker service. This means that
when it bind mounts directories with `docker run`, the source path must
exist on the *hosts* file system.

We've fixed this already for tests, with PYTEST_HOST_TMP, but we had not
actually fixed it for actual using the docker image for real. Before the
previous change to use --mount rather than --volume, a job would
succeed, but the files would be in the wrong location. Now we have that
change, the job fails.

We use a simple if inelegant fix for this: we mount the STORAGE_BASE
directories inside the container at the same path as on the host. This
means that the same config value can be used for job-runner copying
files itself, or for passing the host's docker to mount.

As part of this, I changedthe way we supply these values to being in
one time generated docker/.env file. This is more discoverable, usable
outside of just, and more efficant than before. I chose to use
`docker/medium` and `docker/high` as the volume mount locations, for
simplicity, and also, so the the docker machinery is isolated from the
non-docker venv set up.

This include a basic `just docker/functional-test` command to spin up
a job-runner container and run an actual job on it.