fix: Set MAX_DB_WORKER limit in `local_run`
We allow the user to set the number of concurrent jobs using the
`--concurrency` argument and use this to update the worker limit, but we
didn't update the database worker limit.

The distinction is meaningless for local jobs in any case because these
jobs are just generating dummy data, but it meant that jobs ran with
less parallelism than expected. It also lead to "Waiting on available
database workers" messages appearing in the log output which caused
confusion and anxiety.