Fix pretty broken CI
This was prompted by a bug in setup-python@v5, which means that the
post-clean-up job will fail if you do not actually install anything in
pip, as the cache directory will not be created. This has hit a few of
our GHA workflows.

https://github.com/actions/setup-python/issues/815

The reason this was triggering for job-runner was that the current CI
setup was pretty much broken.

Firstly, the PYTHON_VERSION env var was not being set, which meant that
every test run was always testing 3.8, not 3.9 or 3.10 as per the
matrix.  Fixing this meant a bit of windows/powershell finesse.  It
always explains the fact that for some reason, even though we have 22.04
in the matrix, we did not run the tests on it (this was what triggered
the above failure).

Additionally, we were going to the trouble of installing/starting docker
on macos, and then not even running the docker tests. This feels like we
had the abandonded attempts of getting docker tests running on macos.

To fix this, I just removed those steps and fully embraced not running
docker tests (which required excluding a few more tests).  Additionally,
macos-12 is old, and has very slow runners, so I took the opportunity to
update to using macos-13.

I did attempt to use our opensafely-core/setup-action, but it still
seems to have problems with just on windows, and my yak stack was full.