Add a dedicated manual maintenance mode flag
Currently, manually enabling database maintenance mode is awkward, as by
default it will just get switch off again by the regular automatic
check. So you have to remember to stop that first, which is awkward,
then restart it, for more awkwardness

Instead, this change adds an explicit flag check for
`manual-db-maintenance` - if that is set, we skip the automated check
altogether.

This will be added to the jobrunner justfile in backend-server as
a documented command in a separate PR.