Fix manual maintenance mode
It incorrectly assumed the flag was being set by the caller, and
returned early. The return value is meant to be informative, rather than
acted upon. The tests only tested the return value.

Fixed the function to DTRT, and the tests to check the flag value as
well as the return value

Fixes #762