Do not handle modes if job is to be cancelled.
The db-maintenance and pause modes have special handling at the top of
the handle_job method that eagerly handle those options.

This introduced a bug that job that are to be be cancelled do not get
cancelled if those modes affect them. This means that user's db jobs are
not cancelled when in db-maintenance mode, for example.

This quick fix simply protects the eager mode handling to apply only to
jobs that are not about to be cancelled, which will let the cancellation
logic proceed as normal.

This is probably not the ideal solution, but a) this is affecting users
right now with the extended db-maintenance mode and b) we are about to
completely refactor all the things, so we can live with a less than
ideal solution