Agent only minimally violates layers
The only db writes now happen in the top function, an emulate future API
calls.

 - A GET returns all active jobs from the controller
 - A per-job POST updates the controller with the current state of the
   job

We now only pass a job_definition down to the handle_running_job, not
a Job, and there is no db access in that function any more.

This kicks the tyres on the idea the agent can be db-less