Add `test-docker` to required checks
I'm not sure why this wasn't included in the original list of required
checks. This workflow existed at the point they were [added][1] but
maybe because we weren't using Docker for deployment at the time we
didn't consider it a critical test. In any case, I think we should treat
it as a required pass now.

[1]: https://github.com/opensafely-core/job-runner/commit/45221db1e2b35dcd408bfb35763b990e8b7e4500