Add an initial data schema for the Task API
It is just simple dataclass, and a couple of Enums

The TaskStage enum is intended to be a merger of all valid stages for
all possible TaskTypes. It may well be there is a better way to express
this, but this was the agreed upon initial approach.

This schema is not used in this commit, and is completely declarative,
so has no tests. But is exercised in subsequent commits.