Add an initial implementation of the Task API
This anticipates the agent/controller split, so adds the respective APIs
into new `agent` and `controller` directories.

The controller task API provides a way to insert new tasks, and to mark
them as completed, i.e. no longer active.

The agent task API provides functions to get a list of current tasks,
and to update a task's stage and data.  Right now, the agent task API
writes to the controllers tasks table directly - this is a temporary
hack to allow us to separate the agent and controller out. It will later
be replaced with a HTTP API that the controller will provide and the
agent will call.

Whilst a task is executing, the main purpose of `update_controller` is
to provide fine grained task stage transitions with timing to the
controller to be used for telemetry.  Once the task is completed, the
`update_controller` call should include the results of the task as
a json dict.  That dict is currently unspecified, as that is part of the
split work.

We currently only define one task, RUNJOB.

This is a fairly rough cut API, and we may well want to rework it, but
it should hopefully help us to do the split work faster.