Save job results to disk rather than in memory
Saved all the results to the `metadata.json` file in the job log directory. We were already saving this file now we are adding some information that was missing.

We also fixed the bug in `get_results` where it nonsensically returned a job status when there were no results available. This was never hit in production because of another code path but we have fixed it to correctly raise an error.