Ensure storage dirs are correctly set up for local dev
Updates dotenv-sample to remove the defaults for HIGH_PRIVACY_STORAGE_BASE
and MEDIUM_PRIVACY_STORAGE_BASE. These need to be directories that
exist locally, and were confusingly set to the expected location they
would be mounted into in the container (/workdir). We expect them to use
the workdir/ within the repo for local dev, but they also need to be
absolute paths, so we can't hardcode them into the dotenv-sample.

Instead they are now left empty, and a very simple local_setup script
is added, which creates the .env from dotenv-sample if it doesn't
exist, and replaces the storage lines with absolute paths to the
current workdir/.

local_setup.sh can be invoked on its own, but we also add it into the
just _dotenv command.