Add workaround for job log dir having current month in.
Now that a job's metadata.json is production data, we need to be able to
look it up.

We currently include the current year and month in the directory path,
in order to not just dump them all in one large directory. However, that
date is set at time of writing the file, and we may need to now read it
to respond to the controller, possibly when the month has changed.

So, we add a fallback if we can't find the file where we expect, we'll
glob for it. This is a bit of hack, but it moves us forward. We may want
to revisit how we store logs and results in future.