Refactor update_controller wrapper to take JobStatus
This is much cleaner, as we already have that available and ready.

We use JobStatus.timestamp_ns for the call to task_api. Shortly that
won't be needed, but we will instead want to use that timestamp for
telemtry in future.

Also, mkae JobsStatus.timestamp_ns default to current time, as we
discussed in mob session, for consistency.