Simulate behaviour of future HTTP endpoint
This avoids the need for a _second_ controller loop which marks
tasks as inactive. For RUNJOB tasks we could mark these as inactive as
part of the `handle_job()` function. But that's the wrong place to be
doing this and in any case it won't help with CANCELJOB tasks which have
no equivalent state transition waiting on them.