Switch service.py to using split agent/controller rather than run.py
We may want to rethink our naming a bit in future, too many mains!

Note that running both agent and controller in threads is a temporary
thing - they will be moving to being independant processes soon.

This is just a quick path to being able to run them both simulataneously
using the existing machinery.

Tested manually with add job that it works