Refactor the way we get results from the executor
Previously, we had added get_metadata as a bit of hacky workaround in
order to be able test things for real quickly.

This change is a better implementation of it - we add a results field to
JobStatus.  This will always contain any job metadata that exists on
disk, and makes get_status() the main interface for getting a jobs
results, which is consistant and feels natural.

This makes it work nicely controller asks about a job which is already
completed.

Now, the update_controller has everything it needs in the passed
job status