Change results dict inteface
This previously had two top level keys:
 - `results`, which had the contents of metadata.json
 - `error`, which would containe future error information.

This was based on the assumption that there would be a separate
error.json containing the error information, independantly of
metadata.json.

However, when implenting that, is became clear it would be much easier
to just add error information into metadata.json. The file is still
immutable once written, and as well as error information, it would
include other possibly useful information for metrics or future
inspection.

So, this change makes the results send to the controller just simply the
contents of metadata.json. It still has a top level "error" key with
error information, which the controller uses, but it results information
is now all just in the dictm.