Remove get_results() from ExecutorAPI
Now that we return the results as part getting current JobStatus, we
don't need a separate way to access them.

This removes get_results from ExecutorAPI altogether. This requires now
changes to agent code, just removing the function and various alt
implementations, and fixing tests to use the results dict returned by
get_status()

It means the the agent doesn't import or otherwise rely on JobResults
anymore, it is just a controller concern.