Handle None values in span attributes without logging
We log an error if we try to set span metadata with a type that
otel can't handle. None is one of those values, but it's not very
interesting, and because it happens quite a lot, we end up with
logs full of these errors. Now we just coerce it to a string without
logging it.