Change get_status to handle completed tasks first
If there is a metadata.json, we now the job has been finalized, and is
thus completed.  If there was an error, then we return ERROR, otherwise
we return FINALIZED. This is part of making what we send to the
controller consistent, and is a consquence of having this communication
going over lossy network, and not part of the same sequential loop.

This effects some of the behaviour around cancellation, as reflected in
the test changes.

There are no explicit tests for get_status, as it is so integral that it
gets fully exercised by higher level tests.