Make mark_task_as_error work.
It captures basic exception information, and calls api.finalize with
it.  Just the one pre-existing test working for now, more to follow.

Had to extend StubExecutor to support persisting the error information
passed to.

The fact we had handle_single_job and trace_single_job made things
awkward, and was a controller hang over. Having the exception handling
and tracing unified in the same block is much better, so collapsed them
into one function, which also fix some issues with the test.