Make USING_DUMMY_DATA_BACKEND agent-only config
The controller will become responsible for all backends in future,
so logic about whether the current backend is a dummy data one
belongs with the agent only.

This also removes the logic that set the default to True if the
backend is "expectations"; that backend is no longer active, and we
can depend on just the explicit env var.