Don't treat transient errors in agent as job failures
At present database locking errors are the only ones we consider
transient, but we may want to expand this in future.

Note that the `is_transient_error()` function looks like a duplication
from the controller, but I think the kinds of error we might get in the
agent are much broader and I expect these functions to diverge in
future.