Update executor tests to wait for job to have executed
api.execute() -- and other executor methods -- no longer return a
JobStatus.  Instead, we now call api.get_status() to get a job's status.

In the time between calling api.execute() and api.get_status(), a job
may or may not have finished executing.  So instead of asserting that a
job is in the EXECUTING or EXECUTED state, we now wait for it to have
executed.