Filter jobs by backend in cli commands
Where appropriate, filter jobs by backend in the CLI commands. This
currently uses the agent's BACKEND as a default, in preparations for
at least some of these commands being run by the controller outside of
the backend itself. It does not yet do the work of splitting the
commands between agent and controller and making them runnable by the
controller only (via tasks).