Ensure consistant metadata dict interface
This means we can write code that assumes that all metadata keys will be
present, and don't have to deal with the possiblity of them not being
there.

If not present, they are provided with default value, specified
METADATA_DEFAULTS. This is defined adjacent to the get_job_metadata
function, to aid in keeping these keys in sync.