Handle WAITING_ON_REBOOT in paused mode
When prepare-for-reboot is run from the controller, outside of the
backend, we will need to have paused first. The controller will
continue to run (in order to pick up tasks to kill jobs), but we
don't want jobs that prepare-for-reboot has put back into pending
to be picked up and run.  This change just preserves the "waiting on
reboot" status to differentiate it from jobs that are paused for
some other reason.