Prepare for reboot creates cancel_job tasks
Prepare for reboot now sets the job's status and status code as
before, but now just calls cancel_job(), to deativate runjob
tasks and create cancel_job tasks.

We now expect that the agent is paused (i.e. no new runjob tasks
will be created) but that the agent service will still be running (as
it'll need to pick up the cancel tasks)