Use more of production code in test factory.
We could refactor the producction logic into a call that includes the
transaction, but that means we'd set always the job.status_code to
EXECUTING when creating a task, even in tests.

If the test author has setup the job with a different status_code,
e.g. a FINALIZED job, it would be confusing if it was suddenly set back
to EXECUTING.

But perhaps we should do the work to make this easy and consistant, it
may be worth it.