Store and communicate metrics via job results.
Previously, the sync api had cheated a bit, and when sending updates to
job-server, it just grabbed the latest metrics from the metrics db and
included them. The goal was to communicate live metrics up to job-server
as the job was running, which may enable job-server to improve visiblity
of running jobs.  But that has never really been used, AFAIK.

With the split, the sync thread cannot use the metrics db directly. So,
we need to plumb in the metrics properly.

This simplest way to do this is to record the final metrics (max/mean
cpu and mem usage) for the job in metadata.json when we finalize it. We
already communicate this up to the controller, where sync code can
access it.

The only change here is that we won't stream the current cpu/mem usage
up to job-server - we'll just provide it when the job is completed. This
is argubly less confusing anyway.

Note that this is just user facing metrics - the telemetry still
provides streaming metrics via a separate channel for operators to use.