Drive by: remove broken default values from dotenv-sample
Setting an empty string overrode the defaults, and ended up having
a relative path, which doesn't work, as docker mounts require an
absolute path.

This removes the bad defaults from dotenv-sample, and adds an assert to
prevent manual misconfiguration.

In TPP and test backends, these values are already absolute, this just
hit dev.