Use job ids as the JobDefinition inputs for a task
Instead of retrieving a job's output files and sending them in the
task definition, we now send a list of job ids only. It will be up to
the agent to look up the output files from the job's metadata.json.

This means that the agent also does not need to send any output
filenames out to the controller in results.