Refactor executor methods to take job_id only
We're going to need to read job metadata for other jobs, in order to
identify which files from previous jobs we need to copy into the
container, and for which we only have a job id, not a full job definition.

These methods were being passed the full job definition, but only actually
use the id, so we now just pass the id, which means we can avoid having
to construct a full job definition for dependent jobs.