Tweak telemetry on agent errors
Previously we had logic and tracked transient errors. This change flips
that logic, and assumes transient by default, and provides logic for
detecting fatal errors.

It also flips the telemetry to track fatal_errors attribute, and
*always* sets it, either to true of false. This is a better pattern with
honeycomb than only sending an attribute if something happens, and
omiting it when it doesn't