Make the same changes to error telemetry for the controller.
We now track fatal errors, rather than transient, just like the previous
change to the agent

Additionally, I realise that we were tracking the error on the root LOOP
span, rather that the job specific LOOP_JOB span, due the way
handle_single_job and trace_single_job are nested.

We got rid of this in the agent, so I did the same here, and
consolidated the two into once function, which means we are now
annotating the correct span.