Add trace metadata to metrics spans
We ask the controller to give us the active tasks, and match up the
running containers to the task metadata to annotate the spans with more
detailed information.

A possible rare corner case is that there is a running job container
with no active task, so we default to some fallback attributes that
still give us the minimum we need to filter appropriately.