Fix slow docker container ls calls
Instead of calling it on every metrics tick, do it once per container,
and cache the result, as it won't change.

It also means we can use docker inspect, which does seem to respect the
DOCKER_TZ env var, so it changes our datestr parsing. Well, we have
different but similar hacky workaround.

The cache is self cleaning, but add some telemetry so we can monitor the
cache sise in production.