Add a new INITIATED status code
A job starts off with a status code of CREATED. It needs something
to move to when we create the RUNJOB task, otherwise set_code can't
tell that it's changed and should be traced. It could be moved to
PREPARING, but that's not the same as preparing on the agent.
Plus it might be useful to be able to track the time between the task
being created by the controller and it being moved to preparing by the
agent.

This will change the status code that's posted back to job-server, but
afaict it stores but doesn't use the status code, except to check for
"nonzero_exit" and will accept any string we send.