Make tests error if they need the db fixture
Tests that need a database use the `db` fixture, which patches the
DATABASE_FILE config to a temporary filepath. If you write a test
and forget to add the db fixture, but you have a local db, the test
will pass, but it'll use your (real) local db. This means that it
(a) messes with your local environment and (b) makes the test fail
in CI only, where there's no local db for it to use.

This adds an autouse fixture that always sets the DATABASE_FILE
config variable to None. If a test attempts to use the
database without the db fixture, it will fail with a helpful error
message.