Split read_job_metadata and read_job_task_metadata
We want to read job metadata for a specific task. However, sometimes
we do just want the stored job metadata, irrespective of task - in
particular, when we retrieve the input files that a job needs from
dependent jobs. Those previous jobs will have run with a different
task, so we need to be able to retrieve them without checking on a
task id match.