Update job metadata message for cancelled jobs
The previous wording ("job cancelled by user") was confusing, because
a CANCELJOB task receive by the agent is not necessarily a result of
a job that has been cancelled by a user. It can also be a job that's
been cancelled by db maintenance, or by a prepare-for-reboot.

If a user cancels a job, the Job will have the cancelled flag set
in the db, and the controller will mark it as failed with the
"Cancelled by user" status message.
https://github.com/opensafely-core/job-runner/blob/afe1e4219284dab1ebfef3538774a41cf0e804b3/jobrunner/controller/main.py#L174

So if we're displaying a cancellation message from the job metadata,
it will be for a job that's been cancelled by some non-user requested
means. (Note that for a user-cancelled job, the job metadata will still
record "cancelled by system" - which is accurate - but the job's
status_message will show the right thing to the user)