Run cleanup() for all cancelled jobs
A job in the PREPARED state has no container yet, but it does have
a volume (which is how get_status() identifies it as PREPARED.
https://github.com/opensafely-core/job-runner/blob/4df192219032451622ce99818402a20abc8e21ca/jobrunner/executors/local.py#L305

We therefore also need to run cleanup after cancelling  a job that
is PREPARED; cleanup will try to delete the container, but it ignores
errors if there's no container to remove, and it will delete the
volume.

Since the cleanup will ignore errors from non-existent containers and
volumes, we can just run it always, irrespective of original state.