Add django bash completion in docker containers
Adds the django bash completion for management commands. I've
taken a copy of the completion script as of Django 5.2.1; we could
wget it from the repo during the image build, but it feels safer to
have a fixed copy of it.

Tested with `just docker/run` and `./manage.py <TAB>`