Only use the task timestamp when a job's status code has changed
When a job's status code has changed, we want to use the task's
timestamp to set its updated_at and status_code_updated_at attributes
so we record the accurate time that this status code changed (on the
agent). If we're just updating an ongoing task that's in the same
status, we use the current timestamp.