Remove **attrs from set_code.
The intent of it was to enable setting arbitrary key-value data on the
telemetry. But it was too flexible, and was only used in one place.

That place set active=False, which doesn't do what you might expect - it
just active=False on the job span. At best this is harmless, at worst
confusing.

So remove it, and make the api harded to misuse.