Refactor set_code job tracing
Previously, when we passed an error argument, it did two things:
 - if it was an exception, we'd record the exception on the span
 - if it was truthy, we'd mark the span as an error

These are really separate things. We want to record all exceptions in
telemetry.  But not all exceptions mean its an error that we need to
track or action - that is really down to the status code.

So we rename error to exception, to clarify what it means.  And we only
set the span to error if its an INTERNAL_ERROR.

This honeycombs default error rate graphs and error handling will be
more useful