Enable specific repos to use ehrQL's event level data
Hardcoding the list like this is obviously not particularly lovely, but
there's precendent for it (e.g. in [job-server][1]) and it has certain
advantages in forcing everything through code review and leaving an
audit trail.

The number of repos is expected to remain small, and the feature as a
whole has a limited lifespan: NOD have agreed that after a period of
testing, event-level data access needs to be either made generally
available or canned completely.

[1]: https://github.com/opensafely-core/job-server/pull/3558